Support for /p/[path]
