--- apps/connect/app/p/[id]/loading.tsx
@@ -1,6 +1,0 @@
-import { PageWrapper } from '@connect/components/common/PageWrapper';
-import { ProjectDetailsSkeleton } from '@connect/components/projects/components/ProjectDetailsSkeleton';
-
-export default function Loading() {
-  return <ProjectDetailsSkeleton />;
-}
