--- apps/connect/app/p/[id]/page.tsx
@@ -1,7 +1,0 @@
-import { ProjectDetailsPage } from '@connect/components/projects/[id]/ProjectDetailsPage';
-
-export const dynamic = 'force-dynamic';
-
-export default function ProjectPage({ params }: { params: { id: string } }) {
-  return <ProjectDetailsPage projectId={params.id} />;
-}
