+++ apps/connect/app/p/[path]/loading.tsx
@@ -1,0 +1,6 @@
+import { PageWrapper } from '@connect/components/common/PageWrapper';
+import { ProjectDetailsSkeleton } from '@connect/components/projects/components/ProjectDetailsSkeleton';
+
+export default function Loading() {
+  return <ProjectDetailsSkeleton />;
+}
