+++ apps/connect/components/projects/components/ProjectItem.tsx
@@ -18,7 +18,7 @@ export function ProjectItem({ project }: { project: ProjectsWithMembers[0] }) {
     <Card>
       <CardActionArea
         LinkComponent={Link}
-        href={`/p/${project.id}`}
+        href={`/p/${project.path}`}
         hrefLang='en'
         sx={{ display: 'flex', gap: 2, p: 2, alignItems: 'normal', justifyContent: 'flex-start' }}
       >
