+++ apps/connect/lib/actions/fetchProject.ts
@@ -1,10 +1,14 @@
 import { prisma } from '@charmverse/core/prisma-client';
 import type { StatusAPIResponse } from '@farcaster/auth-kit';
 
-export async function fetchProject(projectId: string) {
-  const project = await prisma.project.findUnique({
+export async function fetchProject({ id, path }: { id?: string; path?: string }) {
+  if (!id && !path) {
+    return null;
+  }
+
+  const project = await prisma.project.findFirst({
     where: {
-      id: projectId
+      OR: [{ id }, { path }]
     },
     select: {
       description: true,
